% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepeatEpigenomeCollection-class.R
\docType{methods}
\name{filterRepRefAcc,RepeatEpigenomeCollection-method}
\alias{filterRepRefAcc,RepeatEpigenomeCollection-method}
\alias{filterRepRefAcc}
\title{filterRepRefAcc-methods}
\usage{
\S4method{filterRepRefAcc}{RepeatEpigenomeCollection}(.Object,
  minReads = getConfigElement("plotRepTree.meth.minReads"))
}
\arguments{
\item{.Object}{\code{\linkS4class{RepeatEpigenomeCollection}} object}

\item{minReads}{threshold for the minimum number of reads required to cover a repeat}
}
\value{
modified \code{\linkS4class{RepeatEpigenomeCollection}} object
}
\description{
Given a \code{\linkS4class{RepeatEpigenomeCollection}} object, remove repeats that do not fulfill
the coverage criteria in any sample
}
\details{
A repeat must fulfill the criteria in all samples in order to be retained
}
\author{
Fabian Mueller
}
