% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepeatEpigenomeCollection-class.R
\docType{methods}
\name{getAnnot,RepeatEpigenomeCollection-method}
\alias{getAnnot,RepeatEpigenomeCollection-method}
\alias{getAnnot}
\title{getAnnot-methods}
\usage{
\S4method{getAnnot}{RepeatEpigenomeCollection}(.Object)
}
\arguments{
\item{.Object}{\code{\linkS4class{RepeatEpigenomeCollection}} object}
}
\value{
Sample annotation as \code{data.frame}
}
\description{
Return the sample sample annotation table
}
\author{
Fabian Mueller
}
