% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{getArgParser}
\alias{getArgParser}
\title{getArgParser}
\usage{
getArgParser()
}
\value{
an \code{ArgumentParser} (see \code{argparse} package) object
}
\description{
Gets an argument parser for command line arguments for the repeat analysis pipeline
}
\examples{
\donttest{
ap <- getArgParser()
ap$print_help()
}
}
\author{
Fabian Mueller
}

