% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepeatEpigenomeCollection-class.R
\docType{methods}
\name{getComparisonInfo,RepeatEpigenomeCollection-method}
\alias{getComparisonInfo,RepeatEpigenomeCollection-method}
\alias{getComparisonInfo}
\title{getComparisonInfo-methods}
\usage{
\S4method{getComparisonInfo}{RepeatEpigenomeCollection}(.Object, annotCol,
  name.grp1, name.grp2, cmpName = NULL, covariateCols = character(0))
}
\arguments{
\item{.Object}{\code{\linkS4class{RepeatEpigenomeCollection}} object}

\item{annotCol}{name of the column in the sample annotation table on which the comparison is based}

\item{name.grp1}{name of the first group in the corresponding column of the sample annotation table}

\item{name.grp2}{name of the second group in the corresponding column of the sample annotation table}

\item{cmpName}{(optional) strings specifying a name for the comparison}

\item{covariateCols}{column names in the sample annotation table of potentially confounding covariates for the analysis}
}
\value{
Comparison information as structured \code{S3} object of class \code{comparisonInfo}
}
\description{
Retrieve values of differential epigenetic quantifications for each RE in each sample
}
\author{
Fabian Mueller
}
