% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepeatEpigenomeCollection-class.R
\docType{methods}
\name{getComparisons,RepeatEpigenomeCollection-method}
\alias{getComparisons,RepeatEpigenomeCollection-method}
\alias{getComparisons}
\title{getComparisons-methods}
\usage{
\S4method{getComparisons}{RepeatEpigenomeCollection}(.Object,
  annotCols = NULL, minGroupSize = 2L, maxGroupCount = 2,
  allPairwise = TRUE, covariateCols = character(0))
}
\arguments{
\item{.Object}{\code{\linkS4class{RepeatEpigenomeCollection}} object}

\item{annotCols}{names of the column in the sample annotation table on which the comparisons are based.
\code{NULL} (default) indicates that all columns will be inspected for compatible comparisons}

\item{minGroupSize}{minimum number of samples to be considered a group}

\item{maxGroupCount}{maximum number of groups in each sample annotation column}

\item{allPairwise}{if more than two groups exist for a sample annotation column, should all pairwise comparisons be conducted (as opposed to one vs all).
currently the only allowed value is \code{TRUE} (default)}

\item{covariateCols}{column names in the sample annotation table of potentially confounding covariates for the analysis}
}
\value{
List of comparison information objects (structured \code{S3} object of class \code{comparisonInfo})
}
\description{
Retrieve a list of \code{comparisonInfo} objects based
}
\author{
Fabian Mueller
}
