% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepeatEpigenomeCollection-class.R
\docType{methods}
\name{getMarks,RepeatEpigenomeCollection-method}
\alias{getMarks,RepeatEpigenomeCollection-method}
\alias{getMarks}
\title{getMarks-methods}
\usage{
\S4method{getMarks}{RepeatEpigenomeCollection}(.Object)
}
\arguments{
\item{.Object}{\code{\linkS4class{RepeatEpigenomeCollection}} object}
}
\value{
Character vector specifying the epigenetic marks contained in the dataset
}
\description{
Return the epigenetic marks contained in the dataset
}
\author{
Fabian Mueller
}
