% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepeatEpigenomeCollection-class.R
\docType{methods}
\name{getRepeatCovg,RepeatEpigenomeCollection-method}
\alias{getRepeatCovg}
\alias{getRepeatCovg,RepeatEpigenomeCollection-method}
\title{getRepeatCovg-methods}
\usage{
\S4method{getRepeatCovg}{RepeatEpigenomeCollection}(.Object, mark,
  dropEmptySamples = FALSE, type = "numReads")
}
\arguments{
\item{.Object}{\code{\linkS4class{RepeatEpigenomeCollection}} object}

\item{mark}{Epigenetic mark for which the coverage information is retrieved}

\item{dropEmptySamples}{Logical indicating whether samples in which the mark is not present should be dropped}

\item{type}{Coverage type. Should be one of the following:
\code{"numReads"}: [default] number of reads covering a RE
\code{"maxCpGcov"}: [\code{DNAmeth} only] maximum number of reads covering a single CpG
\code{"numInstances"}: [\code{DNAmeth} from genome methylation calls only] number of repeat instances in the genome}
}
\value{
A matrix of dimension numRE X numSamples containing coverage values
}
\description{
Retrieve the number of reads covering each RE in each sample
}
\author{
Fabian Mueller
}

