% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepeatEpigenomeCollection-class.R
\docType{methods}
\name{getRepeatScores,RepeatEpigenomeCollection-method}
\alias{getRepeatScores,RepeatEpigenomeCollection-method}
\alias{getRepeatScores}
\title{getRepeatScores-methods}
\usage{
\S4method{getRepeatScores}{RepeatEpigenomeCollection}(.Object, mark,
  dropEmptySamples = FALSE,
  minCpGcov = getConfigElement("meth.minCpGcov"))
}
\arguments{
\item{.Object}{\code{\linkS4class{RepeatEpigenomeCollection}} object}

\item{mark}{Epigenetic mark for which the score information is retrieved}

\item{dropEmptySamples}{Logical indicating whether samples in which the mark is not present should be dropped}

\item{minCpGcov}{[\code{DNAmeth} only] number of reads covering each CpG in order for it to be considered in the
mean methylation level}
}
\value{
A matrix of dimension numRE X numSamples containing scores. Scores are mean methylation levels for DNA methylation and 
        log2 of the fold change for enrichment based methods.
}
\description{
Retrieve values of epigenetic quantifications for each RE in each sample
}
\author{
Fabian Mueller
}
