% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepeatEpigenomeCollection-class.R
\docType{methods}
\name{getRepRef,RepeatEpigenomeCollection-method}
\alias{getRepRef,RepeatEpigenomeCollection-method}
\alias{getRepRef}
\title{getRepRef-methods}
\usage{
\S4method{getRepRef}{RepeatEpigenomeCollection}(.Object)
}
\arguments{
\item{.Object}{\code{\linkS4class{RepeatEpigenomeCollection}} object}
}
\value{
The repeat reference as \code{\linkS4class{RepeatReference}} object
}
\description{
Return the reference repeat object
}
\author{
Fabian Mueller
}
