% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisManager-class.R
\name{getSampleGroups}
\alias{getSampleGroups}
\title{getSampleGroups}
\usage{
getSampleGroups(anno, addAll = FALSE, addIndividual = FALSE)
}
\arguments{
\item{anno}{sample annotation table. must contain sample ids as row names}

\item{addAll}{flag indicating whether a group containing all samples in the annotation table should also be created}

\item{addIndividual}{flag indicating whether a grouping containing a seperate group for each individual sample should be added}
}
\value{
a list containing lists of sample identifiers. The first level list contains lists for each applicable column in the annotation table. On the second level, each list contains the sample ids for each group identified in that column
}
\description{
given a sample annotation table, return the sample groups
}
\details{
gets the annotation from all columns which don't have only unique or only identical values
}
\author{
Fabian Mueller
}

