% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepeatEpigenomeCollection-class.R
\docType{methods}
\name{getSampleMarkTable,RepeatEpigenomeCollection-method}
\alias{getSampleMarkTable,RepeatEpigenomeCollection-method}
\alias{getSampleMarkTable}
\title{getSampleMarkTable-methods}
\usage{
\S4method{getSampleMarkTable}{RepeatEpigenomeCollection}(.Object)
}
\arguments{
\item{.Object}{\code{\linkS4class{RepeatEpigenomeCollection}} object}
}
\value{
Logical matrix of dimension numSamples X numMarks indicating whether a given sample-mark combination is covered
}
\description{
Return a table containing sample-mark combinations covered in the dataset
}
\author{
Fabian Mueller
}
