% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepeatEpigenomeCollection-class.R
\docType{methods}
\name{getSamples,RepeatEpigenomeCollection-method}
\alias{getSamples}
\alias{getSamples,RepeatEpigenomeCollection-method}
\title{getSamples-methods}
\usage{
\S4method{getSamples}{RepeatEpigenomeCollection}(.Object, marks = NULL)
}
\arguments{
\item{.Object}{\code{\linkS4class{RepeatEpigenomeCollection}} object}

\item{marks}{Optionally specify the the marks which a sample needs to cover in order to be retrieved.}
}
\value{
Character vector specifying the sample identifiers
}
\description{
Return the sample indentifiers contained in the dataset
}
\author{
Fabian Mueller
}

