% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{loadConfigFromCommandLine}
\alias{loadConfigFromCommandLine}
\title{loadConfigFromCommandLine}
\usage{
loadConfigFromCommandLine()
}
\value{
a list of parameters for input to the \link{runAnalysis} function
}
\description{
Sets the configuration from command line options
}
\author{
Fabian Mueller
}
