% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logger.cmd.args}
\alias{logger.cmd.args}
\title{logger.cmd.args}
\usage{
logger.cmd.args(cmdArgs)
}
\arguments{
\item{cmdArgs}{command line options as returned by \code{parse_args} of the \code{argparse} package}
}
\value{
nothing of particular interest
}
\description{
Log command line options
}
\author{
Fabian Mueller
}
