% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalizeMatrix}
\alias{normalizeMatrix}
\title{normalizeMatrix}
\usage{
normalizeMatrix(x, method = "standard", ...)
}
\arguments{
\item{x}{a matrix of values to be normalized}

\item{method}{method of normalization. Currently supported are:
\code{none} (no normalization),
\code{standard} (subtract the mean, devide by standard deviation),
\code{scale} (scale to the interval [0,1]) and
\code{quantile} (Quantile normalization)}

\item{...}{arguments passed down to the actual normalization functions}
}
\value{
the normalized matrix
}
\description{
Apply normalizatio to columns of a given matrix
}
\author{
Fabian Mueller
}

