% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotRepeatAlignmentStats}
\alias{plotRepeatAlignmentStats}
\title{plotRepeatAlignmentStats}
\usage{
plotRepeatAlignmentStats(x)
}
\arguments{
\item{x}{filename for the table or data frame containing alignment statistics}
}
\value{
nothing of particular intest
}
\description{
plot the repeat alignment statistic from the table output by the pipeline step "repeatAlignmentStats"
}
\author{
Fabian Mueller
}
