% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{runAnalysis}
\alias{runAnalysis}
\title{runAnalysis}
\usage{
runAnalysis(anaDir, fileTable = NULL, cmdr = NULL, resetToStep = NULL)
}
\arguments{
\item{anaDir}{Analysis directory. epiRepeatR will manage this directory. If it is non-existing a new analysis will be launched
if it is existing, epiRepeatR will try to resume the analysis}

\item{fileTable}{Table containing file paths of the input files and annotation data. Will be ignored, when resuming an analysis.}

\item{cmdr}{A \code{\linkS4class{CommandR}} object. If NULL (default), the analysis is submitted via system calls}

\item{resetToStep}{If resuming an existing analysis or running from an existing analysis directory, reset the analysis
to the beginning of the specified analysis step prior to running the pipeline}
}
\value{
nothing of particular interest
}
\description{
Run the epiRepeatR pipeline. This is the main function.
}
\author{
Fabian Mueller
}

