% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{setConfigElement}
\alias{setConfigElement}
\title{setConfigElement}
\usage{
setConfigElement(name, value)
}
\arguments{
\item{name}{name of the config item}

\item{value}{value of the config item}
}
\value{
nothing of particular interest.
}
\description{
Set a configuration item to a given value
}
\section{Options used by the package}{

\describe{
  \item{\bold{\code{n.processes}}\code{ = 0}}{
       Number of processes used during parallel computations.
  }
  \item{\bold{\code{rscript.exec}}\code{ = "Rscript"}}{
       \code{Rscript executable.}
  }
  \item{\bold{\code{species}}\code{ = "human"}}{
       The organism/taxon the analysis pertains to.
  }
  \item{\bold{\code{refFasta}}\code{ = } MPII default location}{
       The reference fasta file for repeat elements. See vignette for instructions on how they can be obtained.
  }
  \item{\bold{\code{aligner.bs}}\code{ = "bsmap"}}{
       Aligner used for mapping bisulfite sequencing reads to the repeat reference. Currently only "bsmap" is supported.
  }
  \item{\bold{\code{alignment.params.bs}}\code{ = "-g 3 -v 0.2"}}{
       Additional parameters appended during the command line call to the bisulfite aligner.
  }
  \item{\bold{\code{aligner.chip}}\code{ = "chip_bwa"}}{
       Aligner used for mapping ChIP-seq reads to the repeat reference. Currently only "chip_bwa" and "chip_bowtie2" are supported.
  }
  \item{\bold{\code{alignment.params.chip.bwa}}\code{ = "-t 8 -q 20"}}{
       Additional parameters appended during the command line call to the BWA ChIP-seq aligner.
  }
  \item{\bold{\code{alignment.params.chip.bowtie2}}\code{ = "-t 8 -q 20"}}{
       Additional parameters appended during the command line call to the bowtie2 ChIP-seq aligner.
  }
  \item{\bold{\code{aligner.atac}}\code{ = "atac_bowtie2"}}{
       Aligner used for mapping ATAC-seq reads to the repeat reference. Currently only "atac_bowtie2" is supported.
  }
  \item{\bold{\code{alignment.params.atac.bowtie2}}\code{ = "-t 8 -q 20"}}{
       Additional parameters appended during the command line call to the bowtie2 ATAC-seq aligner.
  }
  \item{\bold{\code{samtools.exec}}\code{ = "samtools"}}{
       Location of the samtools executable.
  }
  \item{\bold{\code{chip.mergeInput}}\code{ = FALSE}}{
       Flag indicating whether all bam files containing Input/WCE for ChIP-seq should be merged into one joint Input.
  }
  \item{\bold{\code{tempDir}}\code{ = tempdir()}}{
        Temporary directory for analysis. Specify to be the empty string ("") to create a temp directory in the analysis directory during runAnalysis.
  }
  \item{\bold{\code{inputBam.mappingStatus}}\code{ = "all"}}{
       Which reads should be extracted from the bam files and subsequently mapped to repetitive elements:
       Only reads also mapped in the source bam file ("mapped"), only reads unmapped in the source bam file ("unmapped"),
       or both ("all", default).
  }
  \item{\bold{\code{plotRepTree.dendroMethod}}\code{ = "repeatFamily"}}{
        Method for plotting the repeat subfamily dendrogram. Valid methods include 
                 grouping by repeat subfamily ("repeatFamily"),
                 grouping by hierarchical clustering based on k-mer counts in the repeat sequence (Euclidean distance, complete linkage) ("hierClust"),
                 grouping by hierarchical clustering based on occurrences of terms in the annotation fields of a repeat 
                 ("annotClust"; requires that the repeat references has been annotated from the EMBL format.)
  }
  \item{\bold{\code{plotRepTree.normEnrich}}\code{ = "none"}}{
       Method for normalizing enrichment data before plotting. Currently supported are:
              \code{none} (no normalization),
              \code{standard} (subtract the mean, devide by standard deviation),
              \code{scale} (scale to the interval [0,1]) and
              \code{quantile} (Quantile normalization)
  }
  \item{\bold{\code{plotRepTree.meth.minCpGs}}\code{ = 2}}{
       Minimum number of CpGs to be contained in a repeat element in order to be shown in the resulting methylation tree plots.
  }
  \item{\bold{\code{plotRepTree.meth.minReads}}\code{ = 100}}{
       Minimum number of reads that must match to a given repeat element in order to be shown in the resulting methylation tree plots.
  }
  \item{\bold{\code{annotCols.replicates}}\code{ = NULL}}{
       Column names or indices in the annotation column used for replicate analysis (in exploratory report)
  }
  \item{\bold{\code{meth.minCpGcov}}\code{ = 5}}{
       Minimum number of reads covering a CpG in a repeat element in order to be considered in computing methylation.
  }
  \item{\bold{\code{meth.methCallFormat}}\code{ = "BisSNP"}}{
       Format of input methylation calling files. Can be one of "BisSNP", "EPP"
  }
  \item{\bold{\code{genomeRepeatTrack}}\code{ = NULL}}{
     Path to an RDS file containing a GenomeRepeatTrack object. These files are used to map the sequencing reads instead of the consensus reference.
     Only meaningful if processing input files of type \code{genomeMethCalling} or \code{genomeAlignment}.
  }
  \item{\bold{\code{debug}}\code{ = FALSE}}{
       \code{Logical specifying whether the debug mode is enabled and additional debug-related output should be provided.}
  }
}
}

\author{
Fabian Mueller
}
