% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepeatEpigenomeCollection-class.R
\docType{methods}
\name{filterRepRefMeth,RepeatEpigenomeCollection-method}
\alias{filterRepRefMeth,RepeatEpigenomeCollection-method}
\alias{filterRepRefMeth}
\title{filterRepRefMeth-methods}
\usage{
\S4method{filterRepRefMeth}{RepeatEpigenomeCollection}(.Object,
  minReads = getConfigElement("plotRepTree.meth.minReads"),
  minCpGs = getConfigElement("plotRepTree.meth.minCpGs"),
  minCpGcov = getConfigElement("meth.minCpGcov"), covgType = "numReads")
}
\arguments{
\item{.Object}{\code{\linkS4class{RepeatEpigenomeCollection}} object}

\item{minReads}{threshold for the minimum number of reads required to cover a repeat}

\item{minCpGs}{threshold for the minimum number of CpGs that must be contained in a repeat}

\item{minCpGcov}{threshold specifying that mininum number of reads that must cover at least one
CpG in the repeat in order for it to be retained}

\item{covgType}{type of read coverage to be applied (important for filtering via \code{minReads}).
See \code{\link{getRepeatCovg,RepeatEpigenomeCollection-method}} for possible values.}
}
\value{
modified \code{\linkS4class{RepeatEpigenomeCollection}} object
}
\description{
Given a \code{\linkS4class{RepeatEpigenomeCollection}} object, remove repeats that do not fulfill
the coverage criteria in any sample
}
\details{
A repeat must fulfill the criteria in all samples in order to be retained
}
\author{
Fabian Mueller
}
