% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepeatEpigenomeCollection-class.R
\docType{class}
\name{RepeatEpigenomeCollection-class}
\alias{RepeatEpigenomeCollection}
\alias{RepeatEpigenomeCollection-class}
\title{RepeatEpigenomeCollection Class}
\usage{
RepeatEpigenomeCollection(quantFns, sampleNames, markNames, annot,
  repRef = RepeatReference())
}
\arguments{
\item{quantFns}{Vector of filenames which contains quantification results}
}
\description{
A class for storing quantitative epigenetic data in repetitive elements of the genome

RepeatEpigenomeCollection Constructor
}
\details{
Multiple epigenetic marks are stored and each repetitive element (RE) in a given reference is assigned quantitative
scores for each mark. A dataset of multiple samples can be annotated.
}
\section{Slots}{

\describe{
  \item{\code{annot}}{
      Annotation for the samples in the dataset.
  }
  \item{\code{repRef}}{
      Repeat reference (of class \code{RepeatReference})
  }
  \item{\code{samples}}{
      Samples contained in the dataset.
  }
  \item{\code{marks}}{
      epigenetic marks contained in the dataset.
  }
  \item{\code{markTypes}}{
      Annotation for the samples in the dataset.
  }
  \item{\code{epiQuant}}{
      Data structure for storing the quantifications of the epigenetic marks across marks and samples
  }

}
}

\section{Methods}{

\describe{
  \item{\code{\link{getSamples,RepeatEpigenomeCollection-method}}}{
      Return the sample indentifiers contained in the dataset.
  }
}
}
\author{
Fabian Mueller

Fabian Mueller
}

